// services/sensor/src/models/sensor.model.ts
export interface ISensor {
  sensor_code: string,
  min_temperature: number, // Para deixar a simulação mais próxima da realidade
  max_temperature: number, // Para deixar a simulação mais próxima da realidade
  min_humidity: number, // Para deixar a simulação mais próxima da realidade
  max_humidity: number // Para deixar a simulação mais próxima da realidade
}